tp3Ej7

+++ TP3_EJ7/src/Vagon.java
@@ -9,7 +9,7 @@
  */
 public class Vagon {
     
-    public double pesoMaximo(){
+    public int pesoMaximo(){
         return 0;
     }
     public boolean esLiviano(){
